fat refactoring: furher work on CScape.Core

+++ CScape.Core/Game/Entities/Message/ItemActionMessage.cs
@@ -29,6 +29,6 @@ namespace CScape.Core.Game.Entities.Message
             Index = index;
         }
 
-        public int EventId => MessageId.ItemAction;
+        public int EventId => (int)MessageId.ItemAction;
     }
 }
